Add Phase 0: clippy cleanup, WASM fix, LICENSE, CHANGELOG, README
- Fix ~86 clippy warnings across all crates (0 remaining)
- Add IntoParallelRefMutIterator to maybe_rayon for WASM compatibility
- Fix author name in workspace Cargo.toml
- Add LICENSE-MIT and LICENSE-APACHE (dual license)
- Add CHANGELOG.md (Keep a Changelog format)
- Rewrite README.md (105 algorithms, 9 crates, architecture)
- Add release roadmap and Phase 0 checklist docs
- Add state-of-the-art review (WASM geospatial gap in WOS)
- All 637 tests pass, release build OK, WASM build OK

+++ crates/algorithms/src/landscape/diversity.rs
@@ -150,7 +150,7 @@ pub fn patch_density(raster: &Raster<f64>, params: DiversityParams) -> Result<Ra
             let mut row_data = vec![f64::NAN; cols];
 
             for (col, out) in row_data.iter_mut().enumerate() {
-                let win_size = (2 * params.radius + 1) as usize;
+                let win_size = 2 * params.radius + 1;
                 let mut grid = vec![f64::NAN; win_size * win_size];
                 let mut valid_count = 0usize;
